create default group if need

+++ java-tools/OsmAndServer/src/main/java/net/osmand/server/controllers/user/FavoriteController.java
@@ -22,12 +22,16 @@ import java.io.*;
 import java.util.Map;
 import java.util.zip.GZIPInputStream;
 
+import static net.osmand.router.RouteExporter.OSMAND_ROUTER_V2;
+
 @Controller
 @RequestMapping("/mapapi/fav")
 public class FavoriteController {
     
     public static final String FILE_TYPE_FAVOURITES = "FAVOURITES";
     public static final String FILE_EXT_GPX = ".gpx";
+    private static final String DEFAULT_GROUP_FILE_NAME = "favorites.gpx";
+    private static final String DEFAULT_GROUP_NAME = "favorites";
     public static final String ERROR_WRITING_GPX_MSG = "Error writing gpx!";
     public static final String ERROR_FILE_WAS_CHANGED_MSG = "File was changed!";
     public static final String ERROR_READING_GPX_MSG = "Error reading gpx!";
@@ -65,7 +69,7 @@ public class FavoriteController {
     @ResponseBody
     public ResponseEntity<String> addFav(@RequestBody String data,
                                          @RequestParam String fileName,
-                                         @RequestParam Long updatetime) throws IOException {
+                                         @RequestParam(required = false) Long updatetime) throws IOException {
         PremiumUserDevicesRepository.PremiumUserDevice dev = userdataService.checkValidateUser();
         GPXFile file = createGpxFile(fileName, dev, updatetime);
         if (file != null) {
@@ -73,7 +77,10 @@ public class FavoriteController {
         } else
             throw new OsmAndPublicApiException(UserdataService.ERROR_CODE_FILE_NOT_AVAILABLE,
                     UserdataService.ERROR_MESSAGE_FILE_IS_NOT_AVAILABLE);
-        
+        if (updatetime == null) {
+            PremiumUserFilesRepository.UserFile userGroupFile = userdataService.getLastFileVersion(dev.userid, fileName, FILE_TYPE_FAVOURITES);
+            updatetime = userGroupFile.updatetime.getTime();
+        }
         return updateFavoriteFile(fileName, dev, updatetime, file);
     }
     
@@ -126,7 +133,9 @@ public class FavoriteController {
     
     private void uploadFavoriteFile(File tmpFile, PremiumUserDevicesRepository.PremiumUserDevice dev, String name, Long updatetime) throws IOException {
         userdataService.uploadFile(InternalZipFile.buildFromFile(tmpFile), dev, name, FILE_TYPE_FAVOURITES, System.currentTimeMillis());
-        userdataService.deleteFileVersion(updatetime, dev.userid, name, FILE_TYPE_FAVOURITES, null);
+        if (updatetime != null) {
+            userdataService.deleteFileVersion(updatetime, dev.userid, name, FILE_TYPE_FAVOURITES, null);
+        }
     }
     
     private ResponseEntity<String> updateFavoriteFile(String fileName, PremiumUserDevicesRepository.PremiumUserDevice dev,
@@ -165,9 +174,13 @@ public class FavoriteController {
     private GPXFile createGpxFile(String groupName, PremiumUserDevicesRepository.PremiumUserDevice dev, Long updatetime) throws IOException {
         PremiumUserFilesRepository.UserFile userGroupFile = userdataService.getLastFileVersion(dev.userid, groupName, FILE_TYPE_FAVOURITES);
         if (userGroupFile == null) {
-            throw new OsmAndPublicApiException(UserdataService.ERROR_CODE_FILE_NOT_AVAILABLE, UserdataService.ERROR_MESSAGE_FILE_IS_NOT_AVAILABLE);
+            if (groupName.equals(DEFAULT_GROUP_FILE_NAME)) {
+                userGroupFile = createDefaultGroup(groupName, dev, updatetime);
+            } else {
+                throw new OsmAndPublicApiException(UserdataService.ERROR_CODE_FILE_NOT_AVAILABLE, UserdataService.ERROR_MESSAGE_FILE_IS_NOT_AVAILABLE);
+            }
         }
-        if (userGroupFile.updatetime.getTime() != updatetime) {
+        if (updatetime != null && userGroupFile.updatetime.getTime() != updatetime) {
             throw new OsmAndPublicApiException(UserdataService.ERROR_CODE_FILE_NOT_AVAILABLE, ERROR_FILE_WAS_CHANGED_MSG);
         }
         InputStream in = userGroupFile.data != null ? new ByteArrayInputStream(userGroupFile.data) : userdataService.getInputStream(userGroupFile);
@@ -180,5 +193,13 @@ public class FavoriteController {
             }
         }
         return null;
+    }
+    
+    private PremiumUserFilesRepository.UserFile createDefaultGroup(String groupName, PremiumUserDevicesRepository.PremiumUserDevice dev, Long updatetime) throws IOException {
+        GPXFile file = new GPXFile(OSMAND_ROUTER_V2);
+        file.metadata.name = DEFAULT_GROUP_NAME;
+        File tmpGpx = createTmpGpxFile(file, groupName);
+        uploadFavoriteFile(tmpGpx, dev, groupName, updatetime);
+        return userdataService.getLastFileVersion(dev.userid, groupName, FILE_TYPE_FAVOURITES);
     }
 }
